registar y salida de vehiculos

+++ src/main/java/co/com/ceiba/parqueadero/domain/Vehiculo.java
@@ -1,5 +1,6 @@
 package co.com.ceiba.parqueadero.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Data;
@@ -14,7 +15,7 @@ public class Vehiculo {
 	private Date fechaini ;
 	private String horaini;
 	private Date fechafin ;
-	private double precio;
+	private  BigDecimal  precio;
